Fix a bug where Me First could modify the opponent's next move.
Example: Using Me First with an ability like Aerilate causes the opponent's Normal-type attack to become Flying-type.

This happens because the move object is not cloned when using it with Me First.

Whether this is actually a bug is unknown.
In the case that it's not, it should be made explicit.